Move SupportedModuleLoader into utilities and remove base Loader, we use DI instead
